# docs/superadmin-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "user-agent",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login payload or MFA required",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "access_token": {
                          "type": "string"
                        },
                        "refresh_token": {
                          "type": "string"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            },
                            "isMfaEnabled": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "requiresMfa": {
                          "type": "boolean",
                          "example": true
                        },
                        "tempToken": {
                          "type": "string"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "SuperAdmin login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/mfa/login": {
      "post": {
        "operationId": "MfaController_mfaLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Full access and refresh tokens",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "jwt.token.here",
                    "refresh_token": "refresh.token.here",
                    "user": {
                      "id": "user_123",
                      "email": "admin@superadmin.com",
                      "name": "Super Admin",
                      "role": "SUPER_ADMIN"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid temp token or MFA code"
          }
        },
        "summary": "Complete MFA login with TOTP code",
        "tags": [
          "Auth - MFA"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "New access token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "new.jwt.token.here"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired token, or session revoked"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Logout and revoke current session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/mfa/status": {
      "get": {
        "operationId": "MfaController_getStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get MFA status for current user",
        "tags": [
          "Auth - MFA"
        ]
      }
    },
    "/api/v1/auth/mfa/enroll": {
      "post": {
        "operationId": "MfaController_enroll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Secret and OTPAuth URL for QR code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "MFA already enabled"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Start MFA enrollment - generates secret and QR code URL",
        "tags": [
          "Auth - MFA"
        ]
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "operationId": "MfaController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MFA enabled successfully"
          },
          "400": {
            "description": "Invalid code or no pending enrollment"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Verify TOTP code and enable MFA",
        "tags": [
          "Auth - MFA"
        ]
      }
    },
    "/api/v1/auth/mfa/disable": {
      "post": {
        "operationId": "MfaController_disable",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaDisableDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "MFA disabled successfully"
          },
          "400": {
            "description": "Invalid code or MFA not enabled"
          },
          "401": {
            "description": "Invalid password"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Disable MFA (requires password and TOTP)",
        "tags": [
          "Auth - MFA"
        ]
      }
    },
...
    "/api/v1/superadmin/audit": {
      "get": {
        "operationId": "AuditController_findAll",
        "parameters": [
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "description": "Filter by tenant ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "description": "Filter by event type",
            "schema": {
              "type": "string",
              "enum": [
                "SUPERADMIN_LOGIN",
                "TENANT_CREATED",
                "TENANT_UPDATED",
                "TENANT_ACTIVATED",
                "TENANT_DEACTIVATED",
                "TENANT_SUBSCRIPTION_UPDATED",
                "SUPPORT_TICKET_CREATED",
                "SUPPORT_TICKET_STATUS_CHANGED"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter events created after this ISO date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter events created before this ISO date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated audit events (descending by createdAt)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "audit_123",
                        "tenantId": "tenant_123",
                        "actor": "SUPER_ADMIN:admin@superadmin.com",
                        "eventType": "TENANT_CREATED",
                        "metadata": {
                          "tenantName": "Acme Corp"
                        },
                        "createdAt": "2025-06-01T12:34:56.000Z",
                        "tenant": {
                          "name": "Acme Corp"
                        }
                      }
                    ],
                    "meta": {
                      "page": 1,
                      "pageSize": 50,
                      "total": 100,
                      "totalPages": 2
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List audit log events with pagination",
        "tags": [
          "Audit"
        ]
      }
    },
    "/api/v1/superadmin/audit/export": {
      "get": {
        "operationId": "AuditController_exportCsv",
        "parameters": [
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "description": "Filter by tenant ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "description": "Filter by event type",
            "schema": {
              "type": "string",
              "enum": [
                "SUPERADMIN_LOGIN",
                "TENANT_CREATED",
                "TENANT_UPDATED",
                "TENANT_ACTIVATED",
                "TENANT_DEACTIVATED",
                "TENANT_SUBSCRIPTION_UPDATED",
                "SUPPORT_TICKET_CREATED",
                "SUPPORT_TICKET_STATUS_CHANGED"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter events created after this ISO date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter events created before this ISO date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file download"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Export audit logs as CSV",
        "tags": [
          "Audit"
        ]
      }
    },
...
  },
  "info": {
    "title": "BG Accountability SuperAdmin API",
    "description": "REST API for the BG Accountability Platform - SuperAdmin Services.\n\n## Overview\nThis API provides administrative capabilities for managing the multi-tenant platform.\nAccess is restricted to users with the `SUPER_ADMIN` role.\n\n## Authentication\nAll endpoints (except login) require a valid JWT Bearer token.\n- Login via `POST /api/v1/auth/login`\n- Include token in header: `Authorization: Bearer <token>`\n- Refresh tokens via `POST /api/v1/auth/refresh`\n\n## Multi-Factor Authentication (MFA)\nSuperAdmin accounts require MFA for enhanced security.\n- If MFA is enabled, login returns a `tempToken`\n- Complete MFA via `POST /api/v1/auth/mfa/login` with tempToken + TOTP code\n- MFA can be managed via `/api/v1/auth/mfa/*` endpoints\n\n## Session Management\n- View active sessions: `GET /api/v1/auth/sessions`\n- Revoke sessions: `DELETE /api/v1/auth/sessions/:id`\n\n## Rate Limiting\n- Default: 100 requests per minute per IP\n- Auth endpoints: 10 requests per minute per IP\n\n## Error Format\nAll errors follow: `{ success: false, message: string, code?: string, details?: object }`\n\n## Pagination\nPaginated endpoints use: `{ items: T[], total: number, page: number, pageSize: number }`\n",
    "version": "1.0.0",
    "contact": {
      "name": "BG Accountability",
      "url": "https://bridgegaps.app",
      "email": "support@bridgegaps.app"
    },
    "license": {
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token (SUPER_ADMIN role required)"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@superadmin.com"
          },
          "password": {
            "type": "string",
            "example": "superpassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "MfaLoginDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "6-digit TOTP code",
            "example": "123456"
          },
          "tempToken": {
            "type": "string",
            "description": "Temporary token from initial login"
          }
        },
        "required": [
          "code",
          "tempToken"
        ]
      },
      "MfaStatusDto": {
        "type": "object",
        "properties": {
          "isMfaEnabled": {
            "type": "boolean",
            "description": "Whether MFA is enabled for the user"
          },
          "lastMfaVerifiedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last time MFA was verified"
          }
        },
        "required": [
          "isMfaEnabled"
        ]
      },
      "MfaEnrollResponseDto": {
        "type": "object",
        "properties": {
          "secret": {
            "type": "string",
            "description": "Base32-encoded secret for manual entry"
          },
          "otpauthUrl": {
            "type": "string",
            "description": "OTPAuth URL for QR code generation"
          },
          "label": {
            "type": "string",
            "description": "Display label for authenticator app"
          }
        },
        "required": [
          "secret",
          "otpauthUrl",
          "label"
        ]
      },
      "MfaCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "6-digit TOTP code",
            "example": "123456"
          }
        },
        "required": [
          "code"
        ]
      },
      "MfaDisableDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "6-digit TOTP code",
            "example": "123456"
          },
          "password": {
            "type": "string",
            "description": "Current password for verification"
          }
        },
        "required": [
          "code",
          "password"
        ]
      },
...
    }
  }
}